Split TaskBuilder specs into separate files.
