+++ src/test/scala/mesosphere/mesos/TaskBuilderTest.scala
@@ -10,10 +10,9 @@ import mesosphere.marathon.state.Container.Docker.PortMapping
 import mesosphere.marathon.state.PathId._
 import mesosphere.marathon.state.{ AppDefinition, Container, PathId, Timestamp, _ }
 import mesosphere.marathon.{ MarathonSpec, MarathonTestHelper, Protos }
-import mesosphere.mesos.protos.{ Resource, TaskID, _ }
+import mesosphere.mesos.protos.{ Resource, _ }
 import org.apache.mesos.Protos.ContainerInfo.DockerInfo
 import org.apache.mesos.{ Protos => MesosProtos }
-import org.joda.time.{ DateTime, DateTimeZone }
 import org.scalatest.{ AppendedClues, GivenWhenThen, Matchers, Suites }
 
 import scala.collection.JavaConverters._
@@ -21,10 +20,10 @@ import scala.concurrent.duration._
 import scala.collection.immutable.Seq
 
 class TaskBuilderAllTest extends Suites(
-  new TaskBuilderPortsSuite,
-  new TaskBuilderDockerContainerSuite,
-  new TaskBuilderMesosContainerSuite,
-  new TaskBuilderEnvironmentSuite
+  new TaskBuilderPortsTestSuite,
+  new TaskBuilderDockerContainerTestSuite,
+  new TaskBuilderMesosContainerTestSuite,
+  new TaskBuilderEnvironmentTestSuite
 )
 
 trait TaskBuilderSuiteBase extends UnitTestLike
@@ -133,1091 +132,9 @@ trait TaskBuilderSuiteBase extends UnitTestLike
 
 }
 
-class TaskBuilderPortsSuite extends TaskBuilderSuiteBase {
 
-  import mesosphere.mesos.protos.Implicits._
 
-  "TaskBuilder" when {
 
-    "given a basic offer and an app definition" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 1.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          portDefinitions = PortDefinitions(8080, 8081)
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, Seq[Option[Int]])] = buildIfMatches(offer, appDef)
-      val (taskInfo, taskPorts) = task.get
-
-      "return a defined task" in { task should be('defined) }
-      "???" in { assertTaskInfo(taskInfo, taskPorts, offer) }
-      "return a task without labels" in { taskInfo.hasLabels should be(false) }
-    }
-
-    "given a basic offern and an app defintion with ports and labels" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 1.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          portDefinitions = Seq(
-            PortDefinition(8080, "tcp", Some("http"), Map("VIP" -> "127.0.0.1:8080")),
-            PortDefinition(8081, "tcp", Some("admin"), Map("VIP" -> "127.0.0.1:8081"))
-          )
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, Seq[Option[Int]])] = buildIfMatches(offer, appDef)
-      val (taskInfo, taskPorts) = task.get
-      val discoveryInfo = taskInfo.getDiscovery
-
-      "return a defined task" in { task should be('defined) }
-      "set the discovery info name" in { discoveryInfo.getName should be("frontend.product") }
-      "set discovery info has the correct framework visibility" in {
-        discoveryInfo.getVisibility should be (MesosProtos.DiscoveryInfo.Visibility.FRAMEWORK)
-      }
-      "set the correct port names" in {
-        discoveryInfo.getPorts.getPorts(0).getName should be("http")
-        discoveryInfo.getPorts.getPorts(1).getName should be("admin")
-      }
-      "set correct port protocol" in {
-        discoveryInfo.getPorts.getPorts(0).getProtocol should be("tcp")
-        discoveryInfo.getPorts.getPorts(1).getProtocol should be("tcp")
-      }
-      "set correct port numbers" in {
-        discoveryInfo.getPorts.getPorts(0).getNumber should be(taskPorts(0).get)
-        discoveryInfo.getPorts.getPorts(1).getNumber should be(taskPorts(1).get)
-      }
-      "set correct port labels" in {
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getValue should be("127.0.0.1:8080")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getValue should be("127.0.0.1:8081")
-      }
-
-    }
-
-    "given an offer with an empty port range" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 1.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 31000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          portDefinitions = Seq(
-            PortDefinition(8080, "tcp", Some("http"), Map("VIP" -> "127.0.0.1:8080")),
-            PortDefinition(8081, "tcp", Some("admin"), Map("VIP" -> "127.0.0.1:8081"))
-          )
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, Seq[Option[Int]])] = buildIfMatches(offer, appDef)
-
-      "return an undefined task" in { task should not be ('defined) }
-    }
-
-    "given an app definition with port on tcp and udp" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 1.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          portDefinitions = Seq(
-            PortDefinition(8080, "udp,tcp", Some("http"), Map("VIP" -> "127.0.0.1:8080")),
-            PortDefinition(8081, "udp", Some("admin"), Map("VIP" -> "127.0.0.1:8081"))
-          )
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, Seq[Option[Int]])] = buildIfMatches(offer, appDef)
-      val (taskInfo, taskPorts) = task.get
-      val discoveryInfo = taskInfo.getDiscovery
-
-      "return a defined task" in { task should be('defined) }
-      "set the discovery info name" in { discoveryInfo.getName should be("frontend.product") }
-      "set discovery info has the correct framework visibility" in {
-        discoveryInfo.getVisibility should be (MesosProtos.DiscoveryInfo.Visibility.FRAMEWORK)
-      }
-      "set the correct port names" in {
-        discoveryInfo.getPorts.getPorts(0).getName should be("http")
-        discoveryInfo.getPorts.getPorts(1).getName should be("http")
-        discoveryInfo.getPorts.getPorts(2).getName should be("admin")
-      }
-      "set correct port protocol" in {
-        discoveryInfo.getPorts.getPorts(0).getProtocol should be("udp")
-        discoveryInfo.getPorts.getPorts(1).getProtocol should be("tcp")
-        discoveryInfo.getPorts.getPorts(2).getProtocol should be("udp")
-      }
-      "set correct port numbers" in {
-        discoveryInfo.getPorts.getPorts(0).getNumber should be(taskPorts(0).get)
-        discoveryInfo.getPorts.getPorts(1).getNumber should be(taskPorts(0).get)
-        discoveryInfo.getPorts.getPorts(2).getNumber should be(taskPorts(1).get)
-      }
-      "set correct port labels" in {
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getValue should be("127.0.0.1:8080")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getValue should be("127.0.0.1:8080")
-        discoveryInfo.getPorts.getPorts(2).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(2).getLabels.getLabels(0).getValue should be("127.0.0.1:8081")
-      }
-    }
-
-    "given a basic offer and an app definition with port name, different protocol and labels" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 1.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          portDefinitions = Seq(
-            PortDefinition(8080, "tcp", Some("http"), Map("VIP" -> "127.0.0.1:8080")),
-            PortDefinition(8081, "udp", Some("admin"), Map("VIP" -> "127.0.0.1:8081"))
-          )
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, Seq[Option[Int]])] = buildIfMatches(offer, appDef)
-      val (taskInfo, taskPorts) = task.get
-      val discoveryInfo = taskInfo.getDiscovery
-
-      "return a defined task" in { task should be('defined) }
-      "set the discovery info name" in { discoveryInfo.getName should be("frontend.product") }
-      "set discovery info has the correct framework visibility" in {
-        discoveryInfo.getVisibility should be (MesosProtos.DiscoveryInfo.Visibility.FRAMEWORK)
-      }
-      "set the correct port names" in {
-        discoveryInfo.getPorts.getPorts(0).getName should be("http")
-        discoveryInfo.getPorts.getPorts(1).getName should be("admin")
-      }
-      "set correct port protocol" in {
-        discoveryInfo.getPorts.getPorts(0).getProtocol should be("tcp")
-        discoveryInfo.getPorts.getPorts(1).getProtocol should be("udp")
-      }
-      "set correct port numbers" in {
-        discoveryInfo.getPorts.getPorts(0).getNumber should be(taskPorts(0).get)
-        discoveryInfo.getPorts.getPorts(1).getNumber should be(taskPorts(1).get)
-      }
-      "set correct port labels" in {
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getValue should be("127.0.0.1:8080")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getValue should be("127.0.0.1:8081")
-      }
-    }
-
-    "given an offer and an app definition with port mapping with name, protocol and labels" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 1.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          container = Some(Docker(
-            network = Some(DockerInfo.Network.BRIDGE),
-            portMappings = Some(Seq(
-              PortMapping(
-                containerPort = 8080,
-                hostPort = Some(0),
-                servicePort = 9000,
-                protocol = "tcp",
-                name = Some("http"),
-                labels = Map("VIP" -> "127.0.0.1:8080")
-              ),
-              PortMapping(
-                containerPort = 8081,
-                hostPort = Some(0),
-                servicePort = 9001,
-                protocol = "udp",
-                name = Some("admin"),
-                labels = Map("VIP" -> "127.0.0.1:8081")
-              )
-            ))
-          ))
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, Seq[Option[Int]])] = buildIfMatches(offer, appDef)
-      val (taskInfo, taskPorts) = task.get
-      val discoveryInfo = taskInfo.getDiscovery
-
-      "return a defined task" in { task should be('defined) }
-      "set the discovery info name" in { discoveryInfo.getName should be("frontend.product") }
-      "set discovery info has the correct framework visibility" in {
-        discoveryInfo.getVisibility should be (MesosProtos.DiscoveryInfo.Visibility.FRAMEWORK)
-      }
-      "set the correct port names" in {
-        discoveryInfo.getPorts.getPorts(0).getName should be("http")
-        discoveryInfo.getPorts.getPorts(1).getName should be("admin")
-      }
-      "set correct port protocol" in {
-        discoveryInfo.getPorts.getPorts(0).getProtocol should be("tcp")
-        discoveryInfo.getPorts.getPorts(1).getProtocol should be("udp")
-      }
-      "set correct port numbers" in {
-        discoveryInfo.getPorts.getPorts(0).getNumber should be(taskPorts(0).get)
-        discoveryInfo.getPorts.getPorts(1).getNumber should be(taskPorts(1).get)
-      }
-      "add VIP labels" in {
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getValue should be("127.0.0.1:8080")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getValue should be("127.0.0.1:8081")
-      }
-      "add network-scope labels" in {
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(1).getKey should be("network-scope")
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(1).getValue should be("host")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(1).getKey should be("network-scope")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(1).getValue should be("host")
-      }
-    }
-
-    "given an offer and an app definition with port mapping with name, protocol and labels but no host port" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 1.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          container = Some(Docker(
-            network = Some(DockerInfo.Network.BRIDGE),
-            portMappings = Some(Seq(
-              PortMapping(
-                containerPort = 8080,
-                hostPort = Some(0),
-                servicePort = 9000,
-                protocol = "tcp",
-                name = Some("http"),
-                labels = Map("VIP" -> "127.0.0.1:8080")
-              ),
-              PortMapping(
-                containerPort = 8081,
-                hostPort = None,
-                servicePort = 9001,
-                protocol = "udp",
-                name = Some("admin"),
-                labels = Map("VIP" -> "127.0.0.1:8081")
-              )
-            ))
-          ))
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, Seq[Option[Int]])] = buildIfMatches(offer, appDef)
-      val (taskInfo, taskPorts) = task.get
-      val discoveryInfo = taskInfo.getDiscovery
-
-      "return a defined task" in { task should be('defined) }
-      "set the discovery info name" in { discoveryInfo.getName should be("frontend.product") }
-      "set discovery info has the correct framework visibility" in {
-        discoveryInfo.getVisibility should be (MesosProtos.DiscoveryInfo.Visibility.FRAMEWORK)
-      }
-      "set the correct port names" in {
-        discoveryInfo.getPorts.getPorts(0).getName should be("http")
-        discoveryInfo.getPorts.getPorts(1).getName should be("admin")
-      }
-      "set correct port protocol" in {
-        discoveryInfo.getPorts.getPorts(0).getProtocol should be("tcp")
-        discoveryInfo.getPorts.getPorts(1).getProtocol should be("udp")
-      }
-      "set correct port numbers" in {
-        discoveryInfo.getPorts.getPorts(0).getNumber should be(taskPorts(0).get)
-        discoveryInfo.getPorts.getPorts(1).getNumber should be(8081)
-      }
-      "add VIP labels" in {
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(0).getValue should be("127.0.0.1:8080")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getKey should be("VIP")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(0).getValue should be("127.0.0.1:8081")
-      }
-      "add network-scope labels" in {
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(1).getKey should be("network-scope")
-        discoveryInfo.getPorts.getPorts(0).getLabels.getLabels(1).getValue should be("host")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(1).getKey should be("network-scope")
-        discoveryInfo.getPorts.getPorts(1).getLabels.getLabels(1).getValue should be("container")
-      }
-    }
-
-    "given an offer with duplicated resources and an app definition" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 1.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000)
-        .addResources(ScalarResource("cpus", 1))
-        .addResources(ScalarResource("mem", 128))
-        .addResources(ScalarResource("disk", 2000))
-        .build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          portDefinitions = PortDefinitions(8080, 8081)
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, Seq[Option[Int]])] = buildIfMatches(offer, appDef)
-      val (taskInfo, taskPorts) = task.get
-
-      "return a defined task" in { task should be('defined) }
-      "???" in { assertTaskInfo(taskInfo, taskPorts, offer) }
-      "set no task labels" in { taskInfo.hasLabels should be(false) }
-    }
-
-    "given an offer with enough resources and an app definition" should {
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 2.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          portDefinitions = PortDefinitions(8080, 8081)
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(offer, appDef)
-      val Some((taskInfo, _)) = task
-      def resource(name: String): Resource = taskInfo.getResourcesList.asScala.find(_.getName == name).get
-      val portsResource: Resource = resource("ports")
-
-      "set an appropiate cpu share" in { resource("cpus") should be(ScalarResource("cpus", 1)) }
-      "set an appropiate mem share" in { resource("mem") should be(ScalarResource("mem", 64)) }
-      "set an appropiate disk share" in { resource("disk") should be(ScalarResource("disk", 1)) }
-      "???" in {
-        assert(portsResource.getRanges.getRangeList.asScala.map(range => range.getEnd - range.getBegin + 1).sum == 2)
-      }
-      "set unreserved ports resource role" in { portsResource.getRole should be(ResourceRole.Unreserved) }
-    }
-
-    // #1583 Do not pass zero disk resource shares to Mesos
-    "given an offer and an app definition with zero disk resource" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 2.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          disk = 0.0
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(offer, appDef)
-      val Some((taskInfo, _)) = task
-      def resourceOpt(name: String) = taskInfo.getResourcesList.asScala.find(_.getName == name)
-
-      "return a task with an empty disk resources" in { assert(resourceOpt("disk").isEmpty) }
-    }
-
-    "given an offer and an app definition to share resources" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 2.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000, role = "marathon").build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          portDefinitions = PortDefinitions(8080, 8081)
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(
-        offer,
-        appDef,
-        mesosRole = Some("marathon"),
-        acceptedResourceRoles = Some(Set(ResourceRole.Unreserved, "marathon"))
-      )
-      val Some((taskInfo, _)) = task
-      def resource(name: String): Resource = taskInfo.getResourcesList.asScala.find(_.getName == name).get
-      val portsResource: Resource = resource("ports")
-
-      "set an appropiate cpu share" in { resource("cpus") should be(ScalarResource("cpus", 1, "marathon")) }
-      "set an appropiate mem share" in { resource("mem") should be(ScalarResource("mem", 64, "marathon")) }
-      "set an appropiate disk share" in { resource("disk") should be(ScalarResource("disk", 1, "marathon")) }
-      "???" in {
-        assert(portsResource.getRanges.getRangeList.asScala.map(range => range.getEnd - range.getBegin + 1).sum == 2)
-      }
-      "preserve the resource role" in { portsResource.getRole should be("marathon") }
-    }
-  }
-}
-
-class TaskBuilderDockerContainerSuite extends TaskBuilderSuiteBase {
-
-  import mesosphere.mesos.protos.Implicits._
-
-  "TaskBuilder" when {
-
-    "given an offer and an app definition with a DOCKER container and relative host path" should {
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 2.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 32.0,
-          executor = "//cmd",
-          portDefinitions = Nil,
-          container = Some(Docker(
-            volumes = Seq[Volume](
-              DockerVolume("/container/path", "relativeDirName", MesosProtos.Volume.Mode.RW)
-            )
-          ))
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(offer, appDef)
-      val Some((taskInfo, _)) = task
-      def resource(name: String): Resource = taskInfo.getResourcesList.asScala.find(_.getName == name).get
-
-      // check protobuf construction, should be a ContainerInfo w/ volumes
-      def vol(path: String): Option[MesosProtos.Volume] = {
-        if (taskInfo.hasContainer) {
-          taskInfo.getContainer.getVolumesList.asScala.find(_.getHostPath == path)
-        } else None
-      }
-
-      // sanity, we DID match the offer, right?
-      "set the cpu resource" in {
-        resource("cpus") should be(ScalarResource("cpus", 1))
-      }
-
-      "declare volumes for the container" in {
-        assert(taskInfo.getContainer.getVolumesList.size > 0)
-      }
-      "define a relative directory name" in {
-        vol("relativeDirName") should be('defined) withClue (s"missing expected volume relativeDirName, got instead: ${taskInfo.getContainer.getVolumesList}")
-      }
-    }
-
-    "given an offer and an app definition with a DOCKER container using host-local and external [DockerVolume] volumes" should {
-      import mesosphere.marathon.core.externalvolume.impl.providers.DVDIProviderVolumeToUnifiedMesosVolumeTest._
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 2.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 32.0,
-          executor = "//cmd",
-          portDefinitions = Nil,
-          container = Some(Docker(
-            volumes = Seq[Volume](
-              ExternalVolume("/container/path", ExternalVolumeInfo(
-                name = "namedFoo",
-                provider = "dvdi",
-                options = Map[String, String]("dvdi/driver" -> "bar")
-              ), MesosProtos.Volume.Mode.RW),
-              DockerVolume("/container/path", "relativeDirName", MesosProtos.Volume.Mode.RW)
-            )
-          ))
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(offer, appDef)
-      val Some((taskInfo, _)) = task
-      def resource(name: String): Resource = taskInfo.getResourcesList.asScala.find(_.getName == name).get
-
-      // sanity, we DID match the offer, right?
-      "set the cpu resource" in { resource("cpus") should be(ScalarResource("cpus", 1)) }
-
-      // check protobuf construction, should be a ContainerInfo w/ volumes
-      def vol(name: String): Option[MesosProtos.Volume] = {
-        if (taskInfo.hasContainer) {
-          taskInfo.getContainer.getVolumesList.asScala.find(_.getHostPath == name)
-        } else None
-      }
-
-      "declare container volumes" in { taskInfo.getContainer.getVolumesList.size should be(2) }
-
-      val got1 = taskInfo.getContainer.getVolumes(0)
-      "set container path" in { got1.getContainerPath should be("/container/path") }
-      "set volume mode" in { got1.getMode should be(MesosProtos.Volume.Mode.RW) }
-      "set docker volume name" in { got1.getSource.getDockerVolume.getName should be("namedFoo") }
-      "set docker volume driver" in { got1.getSource.getDockerVolume.getDriver should be("bar") }
-
-      "define the relative directory name" in {
-        vol("relativeDirName") should be('defined) withClue (s"missing expected volume relativeDirName, got instead: ${taskInfo.getContainer.getVolumesList}")
-      }
-    }
-
-    "given an offer and an app definition with a DOCKER container using external [DockerVolume] volumes" should {
-      import mesosphere.marathon.core.externalvolume.impl.providers.DVDIProviderVolumeToUnifiedMesosVolumeTest._
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 2.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 32.0,
-          executor = "//cmd",
-          portDefinitions = Nil,
-          container = Some(Docker(
-            volumes = Seq[Volume](
-              ExternalVolume("/container/path", ExternalVolumeInfo(
-                name = "namedFoo",
-                provider = "dvdi",
-                options = Map[String, String]("dvdi/driver" -> "bar")
-              ), MesosProtos.Volume.Mode.RW),
-              ExternalVolume("/container/path2", ExternalVolumeInfo(
-                name = "namedEdc",
-                provider = "dvdi",
-                options = Map[String, String]("dvdi/driver" -> "ert", "dvdi/boo" -> "baa")
-              ), MesosProtos.Volume.Mode.RO)
-            )
-          ))
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(offer, appDef)
-      val Some((taskInfo, _)) = task
-      def resource(name: String): Resource = taskInfo.getResourcesList.asScala.find(_.getName == name).get
-
-      // sanity, we DID match the offer, right?
-      "set the cpu resource" in { resource("cpus") should be(ScalarResource("cpus", 1)) }
-
-      "declare container volumes" in {
-        taskInfo.getContainer.getVolumesList.size should be(2) withClue (s"check that container has 2 volumes declared, got instead ${taskInfo.getExecutor.getContainer.getVolumesList}")
-      }
-
-      val got1 = taskInfo.getContainer.getVolumes(0)
-      "set container path for first volume" in { got1.getContainerPath should be("/container/path") }
-      "set volume mode for first volume" in { got1.getMode should be(MesosProtos.Volume.Mode.RW) }
-      "set docker volume name for first volume" in { got1.getSource.getDockerVolume.getName should be("namedFoo") }
-      "set docker volume driver for first volume" in { got1.getSource.getDockerVolume.getDriver should be("bar") }
-
-      val got2 = taskInfo.getContainer.getVolumes(1)
-      "set container path for second volume" in { got2.getContainerPath should be("/container/path2") }
-      "set volume mode for second volume" in { got2.getMode should be(MesosProtos.Volume.Mode.RO) }
-      "set docker volume name for second volume" in { got2.getSource.getDockerVolume.getName should be("namedEdc") }
-      "set docker volume driver for second volume" in { got2.getSource.getDockerVolume.getDriver should be("ert") }
-      "set docker volume options" in {
-        got2.getSource.getDockerVolume.getDriverOptions.getParameter(0).getKey should be("boo")
-        got2.getSource.getDockerVolume.getDriverOptions.getParameter(0).getValue should be("baa")
-      }
-    }
-  }
-}
-
-class TaskBuilderMesosContainerSuite extends TaskBuilderSuiteBase {
-
-  import mesosphere.mesos.protos.Implicits._
-
-  "TaskBuilder" when {
-
-    "given an offer and an app definition with a MESOS container using named, external [ExternalVolume] volumes" should {
-
-      import mesosphere.marathon.core.externalvolume.impl.providers.DVDIProviderVolumeToUnifiedMesosVolumeTest._
-
-      val offer = MarathonTestHelper.makeBasicOffer(cpus = 2.0, mem = 128.0, disk = 2000.0, beginPort = 31000, endPort = 32000).build
-      val appDef =
-        AppDefinition(
-          id = "/product/frontend".toPath,
-          cmd = Some("foo"),
-          cpus = 1.0,
-          mem = 32.0,
-          executor = "/qazwsx",
-          portDefinitions = Nil,
-          container = Some(Container.Mesos(
-            volumes = Seq[Volume](
-              ExternalVolume("/container/path", ExternalVolumeInfo(
-                name = "namedFoo",
-                provider = "dvdi",
-                options = Map[String, String]("dvdi/driver" -> "bar")
-              ), MesosProtos.Volume.Mode.RW),
-              ExternalVolume("/container/path2", ExternalVolumeInfo(
-                size = Some(2L),
-                name = "namedEdc",
-                provider = "dvdi",
-                options = Map[String, String]("dvdi/driver" -> "ert")
-              ), MesosProtos.Volume.Mode.RW)
-            )
-          ))
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(offer, appDef)
-      val Some((taskInfo, _)) = task
-      def resource(name: String): Resource = taskInfo.getResourcesList.asScala.find(_.getName == name).get
-
-      // sanity, we DID match the offer, right?
-      "set the cpu resource" in {
-        resource("cpus") should be(ScalarResource("cpus", 1))
-      }
-      "not have a container" in {
-        taskInfo.hasContainer should be(false)
-      }
-      "not have a command" in {
-        taskInfo.hasCommand should be(false)
-      }
-      "have an executor with a MESOS container" in {
-        taskInfo.getExecutor.hasContainer should be(true)
-        taskInfo.getExecutor.getContainer.hasMesos should be(true)
-      }
-      "have a MESOS container with volumes" in {
-        // check protobuf construction, should be a ContainerInfo w/ no volumes, w/ envvar
-        taskInfo.getExecutor.getContainer.getVolumesList.size should be(2) withClue (s"check that container has 2 volumes declared, got instead ${taskInfo.getExecutor.getContainer.getVolumesList}")
-      }
-
-      val got1 = taskInfo.getExecutor.getContainer.getVolumes(0)
-      "set container path for first volume" in { got1.getContainerPath should be("/container/path") }
-      "set volume mode for first volume" in { got1.getMode should be(MesosProtos.Volume.Mode.RW) }
-      "set docker volume name for first volume" in { got1.getSource.getDockerVolume.getName should be("namedFoo") }
-      "set docker volume driver for first volume" in { got1.getSource.getDockerVolume.getDriver should be("bar") }
-
-      val got2 = taskInfo.getExecutor.getContainer.getVolumes(1)
-      "set container path for second volume" in { got2.getContainerPath should be("/container/path2") }
-      "set volume mode for second volume" in { got2.getMode should be(MesosProtos.Volume.Mode.RW) }
-      "set docker volume name for second volume" in { got2.getSource.getDockerVolume.getName should be("namedEdc") }
-      "set docker volume driver for second volume" in { got2.getSource.getDockerVolume.getDriver should be("ert") }
-      "set docker volume options" in {
-        got2.getSource.getDockerVolume.getDriverOptions.getParameter(0).getKey should be("size")
-        got2.getSource.getDockerVolume.getDriverOptions.getParameter(0).getValue should be("2")
-      }
-    }
-
-    "given an offer and an app definition with MESOS Docker container" should {
-
-      val offer = MarathonTestHelper.makeBasicOfferWithRole(cpus = 1.0, mem = 128.0, disk = 1000.0, beginPort = 31000, endPort = 31010, role = ResourceRole.Unreserved)
-        .addResources(RangesResource(Resource.PORTS, Seq(protos.Range(33000, 34000)), "marathon"))
-        .build
-      val appDef =
-        AppDefinition(
-          id = "testApp".toPath,
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          container = Some(Container.MesosDocker(
-            image = "busybox",
-            credential = Some(Container.Credential(
-              principal = "aPrincipal",
-              secret = Some("aSecret")
-            ))
-          )),
-          portDefinitions = Seq.empty,
-          ipAddress = Some(IpAddress(networkName = Some("vnet")))
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(offer, appDef)
-      val (taskInfo, _) = task.get
-
-      "define a task" in { task should be('defined) }
-      "set a container" in { taskInfo.hasContainer should be(true) }
-      "set container type to MESOS" in { taskInfo.getContainer.getType should be (MesosProtos.ContainerInfo.Type.MESOS) }
-      "set a mesos container" in { taskInfo.getContainer.hasMesos should be (true) }
-      "set a mesos container image" in { taskInfo.getContainer.getMesos.hasImage should be (true) }
-      "set the mesos container image type to docker" in { taskInfo.getContainer.getMesos.getImage.getType should be (MesosProtos.Image.Type.DOCKER) }
-      "set the image docker" in { taskInfo.getContainer.getMesos.getImage.hasDocker should be (true) }
-      "set image docker credentials" in {
-        taskInfo.getContainer.getMesos.getImage.getDocker.hasCredential should be(true)
-        taskInfo.getContainer.getMesos.getImage.getDocker.getCredential.getPrincipal should be("aPrincipal")
-        taskInfo.getContainer.getMesos.getImage.getDocker.getCredential.hasSecret should be(true)
-        taskInfo.getContainer.getMesos.getImage.getDocker.getCredential.getSecret should be("aSecret")
-      }
-    }
-
-    "given an offer and an app definition with a MESOS AppC container" should {
-
-      val offer = MarathonTestHelper.makeBasicOfferWithRole(cpus = 1.0, mem = 128.0, disk = 1000.0, beginPort = 31000, endPort = 31010, role = ResourceRole.Unreserved)
-        .addResources(RangesResource(Resource.PORTS, Seq(protos.Range(33000, 34000)), "marathon"))
-        .build
-      val appDef =
-        AppDefinition(
-          id = "testApp".toPath,
-          cpus = 1.0,
-          mem = 64.0,
-          disk = 1.0,
-          executor = "//cmd",
-          container = Some(Container.MesosAppC(
-            image = "anImage",
-            id = Some("sha512-aHashValue"),
-            labels = Map("foo" -> "bar", "test" -> "test")
-          )),
-          portDefinitions = Seq.empty,
-          ipAddress = Some(IpAddress(networkName = Some("vnet")))
-        )
-
-      val task: Option[(MesosProtos.TaskInfo, _)] = buildIfMatches(offer, appDef)
-      val (taskInfo, _) = task.get
-
-      "define a task" in { task should be('defined) }
-      "set a container" in { taskInfo.hasContainer should be(true) }
-      "set container type to MESOS" in { taskInfo.getContainer.getType should be (MesosProtos.ContainerInfo.Type.MESOS) }
-      "set a mesos container" in { taskInfo.getContainer.hasMesos should be (true) }
-      "set a mesos container image" in { taskInfo.getContainer.getMesos.hasImage should be (true) }
-      "set the mesos container image type to APPC" in { taskInfo.getContainer.getMesos.getImage.getType should be (MesosProtos.Image.Type.APPC) }
-      "set the image APPC" in { taskInfo.getContainer.getMesos.getImage.hasAppc should be (true) }
-      "set the AppC id" in {
-        taskInfo.getContainer.getMesos.getImage.getAppc.hasId should be(true)
-        taskInfo.getContainer.getMesos.getImage.getAppc.getId should be("sha512-aHashValue")
-      }
-      "set the Appc labels" in {
-        taskInfo.getContainer.getMesos.getImage.getAppc.hasLabels should be(true)
-        taskInfo.getContainer.getMesos.getImage.getAppc.getLabels.getLabels(0).getKey should be("foo")
-        taskInfo.getContainer.getMesos.getImage.getAppc.getLabels.getLabels(0).getValue should be("bar")
-        taskInfo.getContainer.getMesos.getImage.getAppc.getLabels.getLabels(1).getKey should be("test")
-        taskInfo.getContainer.getMesos.getImage.getAppc.getLabels.getLabels(1).getValue should be("test")
-      }
-    }
-  }
-}
-
-class TaskBuilderEnvironmentSuite extends TaskBuilderSuiteBase {
-
-  "TaskBuilder" when {
-
-    "given host ports 1001 and 1002" should {
-
-      val env = TaskBuilder.portsEnv(Seq(0, 0), Helpers.hostPorts(1001, 1002), Seq(None, None))
-
-      "set env variable PORT" in { env("PORT") should be("1001") }
-      "set env variable PORT0" in { env("PORT0") should be("1001") }
-      "set env variable PORT1" in { env("PORT1") should be("1002") }
-      "not set env variable PORT_1" in { env.keys should not contain ("PORT_0") }
-    }
-
-    "given no ports" should {
-
-      val env = TaskBuilder.portsEnv(Seq(), Seq(), Seq())
-
-      "not set any env variable" in { env should be('empty) }
-    }
-
-    "given protocols" should {
-      val env = TaskBuilder.portsEnv(Seq(0, 0), Helpers.hostPorts(1001, 1002), Seq(Some("http"), Some("https")))
-
-      "set env variable PORT" in { env("PORT") should be("1001") }
-      "set env variable PORT0" in { env("PORT0") should be("1001") }
-      "set env variable PORT1" in { env("PORT1") should be("1002") }
-
-      "set env variable PORT_HTTP" in { env("PORT_HTTP") should be("1001") }
-      "set env variable PORT_HTTPS" in { env("PORT_HTTPS") should be("1002") }
-    }
-
-    "given declared ports" should {
-      val env = TaskBuilder.portsEnv(Seq(80, 8080), Helpers.hostPorts(1001, 1002), Seq(None, None))
-
-      "set env variable PORT" in { env("PORT") should be("1001") }
-      "set env variable PORT0" in { env("PORT0") should be("1001") }
-      "set env variable PORT1" in { env("PORT1") should be("1002") }
-
-      "set env variable PORT_80" in { env("PORT_80") should be("1001") }
-      "set env variable PORT_8080" in { env("PORT_8080") should be("1002") }
-    }
-
-    "given deckared named ports" should {
-      val env = TaskBuilder.portsEnv(Seq(80, 8080, 443), Helpers.hostPorts(1001, 1002, 1003), Seq(Some("http"), None, Some("https")))
-
-      "set env variable PORT" in { env("PORT") should be("1001") }
-      "set env variable PORT0" in { env("PORT0") should be("1001") }
-      "set env variable PORT1" in { env("PORT1") should be("1002") }
-      "set env variable PORT2" in { env("PORT2") should be("1003") }
-
-      "set env variable PORT_80" in { env("PORT_80") should be("1001") }
-      "set env variable PORT_8080" in { env("PORT_8080") should be("1002") }
-      "set env variable PORT_443" in { env("PORT_443") should be("1003") }
-
-      "set env variable PORT_HTTP" in { env("PORT_HTTP") should be("1001") }
-      "set env variable PORT_HTTPS" in { env("PORT_HTTPS") should be("1003") }
-    }
-
-    "an app definition without an task id" should {
-      val version = AppDefinition.VersionInfo.forNewConfig(Timestamp(new DateTime(2015, 2, 3, 12, 30, DateTimeZone.UTC)))
-      val runSpec = AppDefinition(
-        id = PathId("/app"),
-        versionInfo = version
-      )
-      val env = TaskBuilder.taskContextEnv(runSpec = runSpec, taskId = None)
-
-      "not set any env variable" in { env should be('empty) }
-    }
-
-    "given a minimal app definition" should {
-      val version = AppDefinition.VersionInfo.forNewConfig(Timestamp(new DateTime(2015, 2, 3, 12, 30, DateTimeZone.UTC)))
-      val runSpec = AppDefinition(
-        id = PathId("/app"),
-        versionInfo = version
-      )
-      val env = TaskBuilder.taskContextEnv(runSpec = runSpec, taskId = Some(Task.Id("taskId")))
-
-      "set env variable MESOS_TASK_ID" in { env("MESOS_TASK_ID") should be("taskId") }
-      "set env variable MARATHON_APP_ID" in { env("MESOS_APP_ID") should be("/app") }
-      "set env variable MARATHON_APP_VERSION" in { env("MESOS_APP_VERSION") should be("2015-02-03T12:30:00.000Z") }
-      "set env variable MARATHON_APP_RESOURCE_CPUS" in { env("MARATHON_APP_RESOURCE_CPUS") should be(AppDefinition.DefaultCpus.toString) }
-      "set env variable MARATHON_APP_RESOURCE_MEM" in { env("MARATHON_APP_RESOURCE_MEM") should be (AppDefinition.DefaultMem.toString) }
-      "set env variable MARATHON_APP_RESOURCE_DISK" in { env("MARATHON_APP_RESOURCE_DISK") should be(AppDefinition.DefaultDisk.toString) }
-      "set env variable MARATHON_APP_RESOURCE_GPUS" in { env("MARATHON_APP_RESOURCE_GPUS") should be(AppDefinition.DefaultGpus.toString) }
-      "set env variable MARATHON_APP_LABELS" in { env("MARATHON_APP_LABELS")  should be("") }
-    }
-
-    "given an app defintion with all fields defined" should {
-      val version = AppDefinition.VersionInfo.forNewConfig(Timestamp(new DateTime(2015, 2, 3, 12, 30, DateTimeZone.UTC)))
-      val taskId = TaskID("taskId")
-      val runSpec = AppDefinition(
-        id = PathId("/app"),
-        versionInfo = version,
-        container = Some(Docker(
-          image = "myregistry/myimage:version"
-        )),
-        cpus = 10.0,
-        mem = 256.0,
-        disk = 128.0,
-        gpus = 2,
-        labels = Map(
-          "LABEL1" -> "VALUE1",
-          "LABEL2" -> "VALUE2"
-        )
-      )
-      val env = TaskBuilder.taskContextEnv(runSpec = runSpec, Some(Task.Id(taskId)))
-
-      "set env variable MESOS_TASK_ID" in { env("MESOS_TASK_ID") should be("taskId") }
-      "set env variable MARATHON_APP_ID" in { env("MARATHON_APP_ID") should be("/app") }
-      "set env variable MARATHON_APP_VERSION" in { env("MARATHON_APP_VERSION") should be("2015-02-03T12:30:00.000Z") }
-      "set env variable MARATHON_APP_DOCKER_IMAGE" in { env("MARATHON_APP_DOCKER_IMAGE") should be("myregistry/myimage:version") }
-      "set env variable MARATHON_APP_RESOURCE_CPUS" in { env("MARATHON_APP_RESOURCE_CPUS") should be("10.0") }
-      "set env variable MARATHON_APP_RESOURCE_MEM" in { env("MARATHON_APP_RESOURCE_MEM") should be("256.0") }
-      "set env variable MARATHON_APP_RESOURCE_DISK" in { env("MARATHON_APP_RESOURCE_DISK") should be("128.0") }
-      "set env variable MARATHON_APP_RESOURCE_GPUS" in { env("MARATHON_APP_RESOURCE_GPUS") should be("2") }
-      "set env variable MARATHON_APP_LABELS" in { env("MARATHON_APP_LABELS") should be("LABEL1 LABEL2") }
-      "set env variable MARATHON_APP_LABEL_LABEL1" in { env("MARATHON_APP_LABEL_LABEL1") should be("VALUE1") }
-      "set env variable MARATHON_APP_LABEL_LABEL2" in { env("MARATHON_APP_LABEL_LABEL2") should be("VALUE2") }
-    }
-
-    "given unsecure labels" should {
-
-      // will exceed max length for sure
-      val longLabel = "longlabel" * TaskBuilder.maxVariableLength
-      var longValue = "longvalue" * TaskBuilder.maxEnvironmentVarLength
-
-      val runSpec = AppDefinition(
-        labels = Map(
-          "label" -> "VALUE1",
-          "label-with-invalid-chars" -> "VALUE2",
-          "other--label\\--\\a" -> "VALUE3",
-          longLabel -> "value for long label",
-          "label-long" -> longValue
-        )
-      )
-
-      val env = TaskBuilder.taskContextEnv(runSpec = runSpec, Some(Task.Id("taskId")))
-        .filterKeys(_.startsWith("MARATHON_APP_LABEL"))
-
-      "set env variable MARATHON_APP_LABELS" in { env("MARATHON_APP_LABELS") should be("OTHER_LABEL_A LABEL LABEL_WITH_INVALID_CHARS") }
-      "set env variable MARATHON_APP_LABEL_LABEL" in { env("MARATHON_APP_LABEL_LABEL") should be("VALUE1") }
-      "set env variable MARATHON_APP_LABEL_LABEL_WITH_INVALID_CHARS" in { env("MARATHON_APP_LABEL_LABEL_WITH_INVALID_CHARS") should be("VALUE2") }
-      "set env variable MARATHON_APP_LABEL_OTHER_LABEL_A" in { env("MARATHON_APP_LABEL_OTHER_LABEL_A") should be("VALUE3") }
-    }
-
-    "given an app definition with an DOCKER containter image" should {
-      val command =
-        TaskBuilder.commandInfo(
-          runSpec = AppDefinition(
-            id = "/test".toPath,
-            portDefinitions = PortDefinitions(8080, 8081),
-            container = Some(Docker(
-              image = "myregistry/myimage:version"
-            ))
-          ),
-          taskId = Some(Task.Id("task-123")),
-          host = Some("host.mega.corp"),
-          hostPorts = Helpers.hostPorts(1000, 1001),
-          envPrefix = None
-        )
-      val env: Map[String, String] =
-        command.getEnvironment.getVariablesList.asScala.toList.map(v => v.getName -> v.getValue).toMap
-
-      "set env variable MESOS_TASK_ID" in { env("MESOS_TASK_ID") should be("task-123") }
-      "set env variable MARATHON_APP_ID" in { env("MARATHON_APP_ID") should be("/test")}
-      "set env variable MARATHON_APP_VERSION" in { env("MARATHON_APP_VERSION") should be("1970-01-01T00:00:00.000Z") }
-      "set env variable MARATHON_APP_DOCKER_IMAGE" in { env("MARATHON_APP_DOCKER_IMAGE") should be("myregistry/myimage:version") }
-    }
-  }
-
-  "given an app definition with port overrides" should {
-    // why?
-    // see https://github.com/mesosphere/marathon/issues/905
-
-    val command =
-      TaskBuilder.commandInfo(
-        runSpec = AppDefinition(
-          id = "/test".toPath,
-          portDefinitions = PortDefinitions(8080, 8081),
-          env = EnvVarValue(Map(
-            "PORT" -> "1",
-            "PORTS" -> "ports",
-            "PORT0" -> "1",
-            "PORT1" -> "2",
-            "PORT_8080" -> "port8080",
-            "PORT_8081" -> "port8081"
-          ))
-        ),
-        taskId = Some(Task.Id("task-123")),
-        host = Some("host.mega.corp"),
-        hostPorts = Helpers.hostPorts(1000, 1001),
-        envPrefix = None
-      )
-    val env: Map[String, String] =
-      command.getEnvironment.getVariablesList.asScala.toList.map(v => v.getName -> v.getValue).toMap
-
-    "set env variable PORT" in { env("PORT") should be("1") }
-    "set env variable PORTS" in { env("PORTS") should be("ports") }
-    "set env variable PORT0" in { env("PORT0") should be("1") }
-    "set env variable PORT1"in { env("PORT1") should be("2") }
-    "set env variable PORT_8080" in { env("PORT_8080") should be("port8080") }
-    "set env variable PORT_8081" in { env("PORT_8081") should be("port8081") }
-  }
-
-  test("PortsEnvWithOnlyPorts") {
-    val command =
-      TaskBuilder.commandInfo(
-        runSpec = AppDefinition(
-          portDefinitions = PortDefinitions(8080, 8081)
-        ),
-        taskId = Some(Task.Id("task-123")),
-        host = Some("host.mega.corp"),
-        hostPorts = Helpers.hostPorts(1000, 1001),
-        envPrefix = None
-      )
-    val env: Map[String, String] =
-      command.getEnvironment.getVariablesList.asScala.toList.map(v => v.getName -> v.getValue).toMap
-
-    assert("1000" == env("PORT_8080"))
-    assert("1001" == env("PORT_8081"))
-  }
-
-  test("PortsEnvWithCustomPrefix") {
-    val command =
-      TaskBuilder.commandInfo(
-        AppDefinition(
-          portDefinitions = PortDefinitions(8080, 8081)
-        ),
-        Some(Task.Id("task-123")),
-        Some("host.mega.corp"),
-        Helpers.hostPorts(1000, 1001),
-        Some("CUSTOM_PREFIX_")
-      )
-    val env: Map[String, String] =
-      command.getEnvironment.getVariablesList.asScala.toList.map(v => v.getName -> v.getValue).toMap
-
-    assert("1000,1001" == env("CUSTOM_PREFIX_PORTS"))
-
-    assert("1000" == env("CUSTOM_PREFIX_PORT"))
-
-    assert("1000" == env("CUSTOM_PREFIX_PORT0"))
-    assert("1000" == env("CUSTOM_PREFIX_PORT_8080"))
-
-    assert("1001" == env("CUSTOM_PREFIX_PORT1"))
-    assert("1001" == env("CUSTOM_PREFIX_PORT_8081"))
-
-    assert("host.mega.corp" == env("CUSTOM_PREFIX_HOST"))
-
-    assert(Seq("HOST", "PORTS", "PORT0", "PORT1").forall(k => !env.contains(k)))
-    assert(Seq("MESOS_TASK_ID", "MARATHON_APP_ID", "MARATHON_APP_VERSION").forall(env.contains))
-  }
-
-  test("OnlyWhitelistedUnprefixedVariablesWithCustomPrefix") {
-    val command =
-      TaskBuilder.commandInfo(
-        AppDefinition(
-          portDefinitions = PortDefinitions(8080, 8081)
-        ),
-        Some(Task.Id("task-123")),
-        Some("host.mega.corp"),
-        Helpers.hostPorts(1000, 1001),
-        Some("P_")
-      )
-    val env: Map[String, String] =
-      command.getEnvironment.getVariablesList.asScala.toList.map(v => v.getName -> v.getValue).toMap
-
-    val nonPrefixedEnvVars = env.filterKeys(!_.startsWith("P_"))
-
-    val whiteList = Seq("MESOS_TASK_ID", "MARATHON_APP_ID", "MARATHON_APP_VERSION", "MARATHON_APP_RESOURCE_CPUS",
-      "MARATHON_APP_RESOURCE_MEM", "MARATHON_APP_RESOURCE_DISK", "MARATHON_APP_RESOURCE_GPUS", "MARATHON_APP_LABELS")
-
-    assert(nonPrefixedEnvVars.keySet.forall(whiteList.contains))
-  }
-
-  test("PortsEnvWithOnlyMappings") {
-    val command =
-      TaskBuilder.commandInfo(
-        runSpec = AppDefinition(
-          container = Some(Docker(
-            network = Some(DockerInfo.Network.BRIDGE),
-            portMappings = Some(Seq(
-              PortMapping(containerPort = 8080, hostPort = Some(0), servicePort = 9000, protocol = "tcp", name = Some("http")),
-              PortMapping(containerPort = 8081, hostPort = Some(0), servicePort = 9000, protocol = "tcp", name = Some("jabber"))
-            ))
-          ))
-        ),
-        taskId = Some(Task.Id("task-123")),
-        host = Some("host.mega.corp"),
-        hostPorts = Helpers.hostPorts(1000, 1001),
-        envPrefix = None
-      )
-    val env: Map[String, String] =
-      command.getEnvironment.getVariablesList.asScala.toList.map(v => v.getName -> v.getValue).toMap
-
-    assert("1000" == env("PORT_8080"))
-    assert("1001" == env("PORT_8081"))
-    assert("1000" == env("PORT_HTTP"))
-    assert("1001" == env("PORT_JABBER"))
-  }
-
-  test("PortsEnvWithBothPortsAndMappings") {
-    val command =
-      TaskBuilder.commandInfo(
-        runSpec = AppDefinition(
-          portDefinitions = PortDefinitions(22, 23),
-          container = Some(Docker(
-            network = Some(DockerInfo.Network.BRIDGE),
-            portMappings = Some(Seq(
-              PortMapping(containerPort = 8080, hostPort = Some(0), servicePort = 9000, protocol = "tcp"),
-              PortMapping(containerPort = 8081, hostPort = Some(0), servicePort = 9000, protocol = "tcp")
-            ))
-          ))
-        ),
-        taskId = Some(Task.Id("task-123")),
-        host = Some("host.mega.corp"),
-        hostPorts = Helpers.hostPorts(1000, 1001),
-        envPrefix = None
-      )
-    val env: Map[String, String] =
-      command.getEnvironment.getVariablesList.asScala.toList.map(v => v.getName -> v.getValue).toMap
-
-    assert("1000" == env("PORT_8080"))
-    assert("1001" == env("PORT_8081"))
-
-    assert(!env.contains("PORT_22"))
-    assert(!env.contains("PORT_23"))
-  }
-}
 
 class TaskBuilderTest extends MarathonSpec
     with AppendedClues
@@ -1921,7 +838,6 @@ class TaskBuilderTest extends MarathonSpec
       .build
     shouldBuildTask("Should take offer with spark:enabled", offerHostB)
   }
-
 
   test("TaskWillCopyFetchIntoCommand") {
     val command = TaskBuilder.commandInfo(
